+++ data-raw/wd_pcon_lad_la_rgn_ctry.R
@@ -1,6 +1,7 @@
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # Create the wd_pcon_lad_la_reg_ctry data set
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# NOTE: API queries in this script may take a couple of minutes to run
 #
 # This data is created by joining wd_pcon_lad_la
 # - https://geoportal.statistics.gov.uk/search?tags=lup_wd_pcon_lad_utla
@@ -125,13 +126,13 @@ wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn_ctry %>%
 # Add 3 digit local authority codes from GIAS  ----------------------------
 wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn_ctry %>%
   # join the data onto the GIAs LA 3 digit code data
-  dplyr::left_join(gias_3_digit_la_codes, by = c(
+  dplyr::left_join(old_la_codes, by = c(
     "la_name" = "la_name",
     "new_la_code" = "new_la_code",
     "old_la_code" = "old_la_code"
   )) %>%
   dplyr::mutate(old_la_code = dplyr::if_else(is.na(old_la_code),
-    "z", old_la_code
+                                             "z", old_la_code
   )) %>%
   dplyr::distinct()