+++ R/zzz.R
@@ -1,0 +1,3 @@
+.onAttach <- function(libname, pkgname) {
+  packageStartupMessage("Note: The dataset `wd_pcon_lad_la_rgn_ctry` is deprecated and will be removed in the next major release. Use `geo_hierarchy` instead.")
+}